// apps/shared/types/units.types.ts
export interface HeroTemplate {
  code: string;
  name: string;
  description: string;
  classCode: string;
  avatar: string;
  level: number;
  combat: number;
  speed: number;
  focus: number;
  resistance: number;
  will: number;
  vitality: number;
  initialSkills: string[];
  initialSpells: string[];
  recruitCost: {
    ore?: number;
    supplies?: number;
    arcane?: number;
    devotion?: number;
  };
  icon: string;
  themeColor: string;
}